Handle “undefined” as passed option values (#146)
Allow the consumers to pass “undefined” as options values instead of misbehaving.

Today if the consumer passes “undefined”, we ignore the default options and allow the “undefined” value to fall through to the rest of the program, resulting in undefined behavior. For an example, the maxDelta becomes NaN if threshold: undefined is provided.

After this change we ensure that “undefined” as options values is converted into a default value.

It is customary in JavaScript, especially in options objects, to treat a missing key and a key with value of undefined the same way. In particular, it is a distinction that is easily introduced when the customer receives a value from somewhere else and wants to continue to support default values.